Add files via upload
“NU180” is a project to create a reverse-engineered drop-in replacement for the U180 hybrid in the Hewlett-Packard / Agilent / Keysight 3458A 8.5 digit DVM. 

The U180 hybrid was designed by HP in the 1980’s. In one package, it includes a custom current-steering chip and a custom resistor array. As many with an 3458A knows all too painfully, its “A3” ADC board, and in particular its custom hybrid U180 within that board, are frequently found to have problems with excessive drift, even to the point that the instrument will fail to function at all. New A3 boards and U180's are not available, and an official repair, if offered at all, is extremely expensive. 

The parts needed for this attempt includes a small amount of control logic, analog switches for current-steering, and a set of well-matched resistors. Here, the control circuitry is managed by a 5V CPLD (ATF1502AS), and the current-steering is implemented with ADG734 SPDT analog switches. A number of high-precision resistors are needed, with particular attention paid to their temperature coefficient of resistance (TCR) values. To stand a chance at achieving the same performance as the original U180, it's expected that the best available resistors are sourced, i.e. with as low as 0.2 ppm TCR, especially for the 80k, 20k and 50k resistors. Many of these are expensive and/or hard to get, and some especially hard to find ones like the 80k's may even need to be sourced in semi-custom form. Therefore, it's expected that the first iterations are made with less expensive parts (e.g. 2 ppm) until functionality is proven. 

Users of the layout may need, at least initially, to change the largest resistors (the 80 and 50k ones, in 2512 English footprints) to something like 1206's, depending on what is available to you. 

Please provide feedback if you find any problems with the board design, or if you have any improvements to offer. Thank you!